chore: extract callbacks package
