+++ api/pybot/callbacks/streaming.py
@@ -1,13 +1,11 @@
-from typing import Any, Optional, Union
+from typing import Any, Optional
 from uuid import UUID
 
 from fastapi import WebSocket
 from langchain.callbacks.base import AsyncCallbackHandler
 from langchain.schema import LLMResult
 
-from pybot.models import Conversation
 from pybot.schemas import ChatMessage
-from pybot.utils import utcnow
 
 
 class StreamingLLMCallbackHandler(AsyncCallbackHandler):
@@ -91,22 +89,3 @@ class StreamingLLMCallbackHandler(AsyncCallbackHandler):
             type="error",
         )
         await self.websocket.send_text(message.model_dump_json())
-
-
-class UpdateConversationCallbackHandler(AsyncCallbackHandler):
-    def __init__(self, conversation_id: str):
-        self.conversation_id: str = conversation_id
-
-    async def on_chain_end(
-        self,
-        outputs: dict[str, Any],
-        *,
-        run_id: UUID,
-        parent_run_id: Optional[UUID] = None,
-        tags: Optional[list[str]] = None,
-        **kwargs: Any,
-    ) -> None:
-        """Run when chain ends running."""
-        conv = await Conversation.get(self.conversation_id)
-        conv.updated_at = utcnow()
-        await conv.save()
